Update JSONs and refactor URL checking
