+++ packages/dd-trace/test/plugins/util/ci-env/buddy.json
@@ -4,7 +4,7 @@
       "BUDDY": "true",
       "BUDDY_EXECUTION_BRANCH": "master",
       "BUDDY_EXECUTION_ID": "buddy-execution-id",
-      "BUDDY_EXECUTION_REVISION": "e5e13f8b7f8d5c6096a0501dc09b48eef05fea96",
+      "BUDDY_EXECUTION_REVISION": "b9f0fb3fdbb94c9d24b2c75b49663122a529e123",
       "BUDDY_EXECUTION_REVISION_COMMITTER_EMAIL": "mikebenson@buddy.works",
       "BUDDY_EXECUTION_REVISION_COMMITTER_NAME": "Mike Benson",
       "BUDDY_EXECUTION_REVISION_MESSAGE": "Create buddy.yml",
@@ -12,7 +12,7 @@
       "BUDDY_EXECUTION_URL": "https://app.buddy.works/myworkspace/my-project/pipelines/pipeline/456/execution/5d9dc42c422f5a268b389d08",
       "BUDDY_PIPELINE_ID": "456",
       "BUDDY_PIPELINE_NAME": "Deploy to Production",
-      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project"
+      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project.git"
     },
     {
       "ci.pipeline.id": "456/buddy-execution-id",
@@ -24,8 +24,8 @@
       "git.commit.committer.email": "mikebenson@buddy.works",
       "git.commit.committer.name": "Mike Benson",
       "git.commit.message": "Create buddy.yml",
-      "git.commit.sha": "e5e13f8b7f8d5c6096a0501dc09b48eef05fea96",
-      "git.repository_url": "https://github.com/buddyworks/my-project",
+      "git.commit.sha": "b9f0fb3fdbb94c9d24b2c75b49663122a529e123",
+      "git.repository_url": "https://github.com/buddyworks/my-project.git",
       "git.tag": "v1.0"
     }
   ],
@@ -34,7 +34,7 @@
       "BUDDY": "true",
       "BUDDY_EXECUTION_BRANCH": "my-name-is-rotag/fix-original-bug",
       "BUDDY_EXECUTION_ID": "buddy-execution-id",
-      "BUDDY_EXECUTION_REVISION": "e5e13f8b7f8d5c6096a0501dc09b48eef05fea96",
+      "BUDDY_EXECUTION_REVISION": "b9f0fb3fdbb94c9d24b2c75b49663122a529e123",
       "BUDDY_EXECUTION_REVISION_COMMITTER_EMAIL": "mikebenson@buddy.works",
       "BUDDY_EXECUTION_REVISION_COMMITTER_NAME": "Mike Benson",
       "BUDDY_EXECUTION_REVISION_MESSAGE": "Create buddy.yml",
@@ -42,7 +42,7 @@
       "BUDDY_EXECUTION_URL": "https://app.buddy.works/myworkspace/my-project/pipelines/pipeline/456/execution/5d9dc42c422f5a268b389d08",
       "BUDDY_PIPELINE_ID": "456",
       "BUDDY_PIPELINE_NAME": "Deploy to Production",
-      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project"
+      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project.git"
     },
     {
       "ci.pipeline.id": "456/buddy-execution-id",
@@ -54,8 +54,8 @@
       "git.commit.committer.email": "mikebenson@buddy.works",
       "git.commit.committer.name": "Mike Benson",
       "git.commit.message": "Create buddy.yml",
-      "git.commit.sha": "e5e13f8b7f8d5c6096a0501dc09b48eef05fea96",
-      "git.repository_url": "https://github.com/buddyworks/my-project",
+      "git.commit.sha": "b9f0fb3fdbb94c9d24b2c75b49663122a529e123",
+      "git.repository_url": "https://github.com/buddyworks/my-project.git",
       "git.tag": "v1.0"
     }
   ],
@@ -64,7 +64,7 @@
       "BUDDY": "true",
       "BUDDY_EXECUTION_BRANCH": "refs/heads/feature/one",
       "BUDDY_EXECUTION_ID": "buddy-execution-id",
-      "BUDDY_EXECUTION_REVISION": "e5e13f8b7f8d5c6096a0501dc09b48eef05fea96",
+      "BUDDY_EXECUTION_REVISION": "b9f0fb3fdbb94c9d24b2c75b49663122a529e123",
       "BUDDY_EXECUTION_REVISION_COMMITTER_EMAIL": "mikebenson@buddy.works",
       "BUDDY_EXECUTION_REVISION_COMMITTER_NAME": "Mike Benson",
       "BUDDY_EXECUTION_REVISION_MESSAGE": "Create buddy.yml",
@@ -72,7 +72,7 @@
       "BUDDY_EXECUTION_URL": "https://app.buddy.works/myworkspace/my-project/pipelines/pipeline/456/execution/5d9dc42c422f5a268b389d08",
       "BUDDY_PIPELINE_ID": "456",
       "BUDDY_PIPELINE_NAME": "Deploy to Production",
-      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project"
+      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project.git"
     },
     {
       "ci.pipeline.id": "456/buddy-execution-id",
@@ -84,8 +84,8 @@
       "git.commit.committer.email": "mikebenson@buddy.works",
       "git.commit.committer.name": "Mike Benson",
       "git.commit.message": "Create buddy.yml",
-      "git.commit.sha": "e5e13f8b7f8d5c6096a0501dc09b48eef05fea96",
-      "git.repository_url": "https://github.com/buddyworks/my-project",
+      "git.commit.sha": "b9f0fb3fdbb94c9d24b2c75b49663122a529e123",
+      "git.repository_url": "https://github.com/buddyworks/my-project.git",
       "git.tag": "0.2.0"
     }
   ],
@@ -94,7 +94,7 @@
       "BUDDY": "true",
       "BUDDY_EXECUTION_BRANCH": "master",
       "BUDDY_EXECUTION_ID": "buddy-execution-id",
-      "BUDDY_EXECUTION_REVISION": "e5e13f8b7f8d5c6096a0501dc09b48eef05fea96",
+      "BUDDY_EXECUTION_REVISION": "b9f0fb3fdbb94c9d24b2c75b49663122a529e123",
       "BUDDY_EXECUTION_REVISION_COMMITTER_EMAIL": "mikebenson@buddy.works",
       "BUDDY_EXECUTION_REVISION_COMMITTER_NAME": "Mike Benson",
       "BUDDY_EXECUTION_REVISION_MESSAGE": "Create buddy.yml",
@@ -102,7 +102,7 @@
       "BUDDY_EXECUTION_URL": "https://app.buddy.works/myworkspace/my-project/pipelines/pipeline/456/execution/5d9dc42c422f5a268b389d08",
       "BUDDY_PIPELINE_ID": "456",
       "BUDDY_PIPELINE_NAME": "Deploy to Production",
-      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project",
+      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project.git",
       "DD_GIT_BRANCH": "user-supplied-branch",
       "DD_GIT_COMMIT_AUTHOR_DATE": "usersupplied-authordate",
       "DD_GIT_COMMIT_AUTHOR_EMAIL": "usersupplied-authoremail",
@@ -139,7 +139,7 @@
       "BUDDY": "true",
       "BUDDY_EXECUTION_BRANCH": "master",
       "BUDDY_EXECUTION_ID": "buddy-execution-id",
-      "BUDDY_EXECUTION_REVISION": "e5e13f8b7f8d5c6096a0501dc09b48eef05fea96",
+      "BUDDY_EXECUTION_REVISION": "b9f0fb3fdbb94c9d24b2c75b49663122a529e123",
       "BUDDY_EXECUTION_REVISION_COMMITTER_EMAIL": "mikebenson@buddy.works",
       "BUDDY_EXECUTION_REVISION_COMMITTER_NAME": "Mike Benson",
       "BUDDY_EXECUTION_REVISION_MESSAGE": "Create buddy.yml",
@@ -147,7 +147,7 @@
       "BUDDY_EXECUTION_URL": "https://app.buddy.works/myworkspace/my-project/pipelines/pipeline/456/execution/5d9dc42c422f5a268b389d08",
       "BUDDY_PIPELINE_ID": "456",
       "BUDDY_PIPELINE_NAME": "Deploy to Production",
-      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project",
+      "BUDDY_SCM_URL": "https://github.com/buddyworks/my-project.git",
       "DD_GIT_BRANCH": "user-supplied-branch",
       "DD_GIT_COMMIT_AUTHOR_DATE": "usersupplied-authordate",
       "DD_GIT_COMMIT_AUTHOR_EMAIL": "usersupplied-authoremail",
